+++ Changelog.md
@@ -3,6 +3,7 @@ Changelog
 
 # Version 1.5.6
 - Improve GBBasic 1.2 detection
+- Detect DevSound X2
 
 # Version 1.5.5
 - Improve GBBasic 1.1 detection (more reliable)